REMOVE: remove network log system. The network environment is very complex. so this system have too problem in error case. if need this function, can use the new feature CURSTOM FUNCTION. ADD: Add new enum for LogOperation, use this opt param can control log output type. such as output to file, or std out, or custom function. CHANGE: Initialize function remove two params, bDebug and bCoverPrev. add a new param emOperation as new LOGOPT type.

+++ univ/Base64.cpp
@@ -32,7 +32,7 @@ char base64decode_lut[] = {	0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,	0,  0
 string CBase64::Encode(const string& strEncode)
 {
 	auto src = (const unsigned char*)strEncode.data();
-	int len = strEncode.length();
+	size_t len = strEncode.length();
 	unsigned char *out, *pos;
     const unsigned char *end, *in;
 
@@ -79,7 +79,7 @@ string CBase64::Encode(const string& strEncode)
 string CBase64::Decode(const string& strDecode)
 {
 	unsigned char* p = (unsigned char*)strDecode.data();
-	int len = strDecode.length();
+	size_t len = strDecode.length();
     int pad = len > 0 && (len % 4 || p[len - 1] == '=');
     const size_t L = ((len + 3) / 4 - pad) * 4;
     std::string str(L / 4 * 3 + pad, '\0');
